+++ courses/cunix2/libft/src/ft_memcmp.c
@@ -1,9 +1,11 @@
 #include "../libft.h"
+
 int ft_memcmp(const void *s1, const void *s2, size_t n)
 {
     const char *s1_str = (const char *)s1;
     const char *s2_str = (const char *)s2;
-    for (size_t i = 0; i < n; i++)
+
+    for (unsigned long i = 0; i < n; i++)
     {
         if (s1_str[i] < s2_str[i])
         {
@@ -14,5 +16,6 @@ int ft_memcmp(const void *s1, const void *s2, size_t n)
             return -1;
         }
     }
+
     return 0;
 }
